+++ Changelog.md
@@ -12,14 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Groq APIに対応。メインエディタ・AIチャット・アイデアプロセッサのいずれでも使用可能  
 - エクスポートウィンドウのDOCX出力機能を実装  
  - ※ルビは変換されません。  
-- OpenCode連携機能を実装。Ctrl+Shift+KでOpenCodeを開くことが可能に  
-
+- OpenCode連携機能を実装。OpenCodeがグローバルインストールされている場合、Ctrl+Shift+Kでopencode serveを起動、別ウィンドウで開くことが可能に  
 
 ### Changed (変更点)    
 
 
 ### Fixed (修正点) 
-
+- 一部のウィンドウに<html lang=ja>が抜けていたので追加  
+- フォントの指定を修正  
 
 
 ## [1.4.1] - 2026-3-16